chore: убрать комментирование
