+++ src/components/App/App.css
@@ -5,7 +5,6 @@
   color: rgba(255, 255, 255, 1);
   background-color: #202020;
   height: 100vh;
-  /*padding: 0 70px;*/
 }
 
 .App:has(.menu_opened) {
