+++ src/components/Main/components/Promo/Promo.css
@@ -46,9 +46,3 @@
     }
 }
 
-/*@media (min-width: 1280px) {*/
-/*    .promo {*/
-/*        padding-left: 140px;*/
-/*        padding-right: 140px;*/
-/*    }*/
-/*}*/
