+++ src/components/Main/components/Promo/Promo.jsx
@@ -1,10 +1,9 @@
 import React from 'react';
 import './Promo.css';
-import Header from '../../../Header/Header';
+
 const Promo = () => {
   return (
     <div className="promo">
-      {/*<Header landing />*/}
       <h1 className="promo__title">Учебный проект студента факультета Веб-разработки.</h1>
     </div>
   );
