+++ src/components/Movies/Movies.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import './Movies.css';
-import Header from '../Header/Header';
 import SearchForm from './SearchForm/SearchForm';
 import MoviesCardList from './MoviesCardList/MoviesCardList';
 import Footer from '../Footer/Footer';
