+++ src/components/Movies/MoviesCard/MoviesCard.css
@@ -34,16 +34,11 @@
 
 }
 .movies__card {
-    /*max-width: 270px;*/
     width: 100%;
-    /*max-height: 151px;*/
 }
 .movies__image {
     max-width: 100%;
     display: block;
-    /*max-height: 151px;*/
-
-    /*max-height: 151px;*/
 }
 .movies__description {
     position: relative;
