+++ src/components/Movies/MoviesCardList/MoviesCardList.css
@@ -15,7 +15,6 @@
 
 @media (max-width: 1279px) {
     .movies__list {
-        /*padding-top: 80px;*/
         grid-template-columns: repeat(2, 1fr);
         column-gap: 30px;
         row-gap: 30px;
