+++ src/components/Movies/SearchForm/SearchForm.css
@@ -1,6 +1,5 @@
 .movies__searchform {
     position: relative;
-    /*padding-top: 70px;*/
     width: 100%;
     margin-bottom: 150px;
 }
@@ -30,12 +29,6 @@
         margin-bottom: 160px;
     }
 }
-
-/*@media (max-width: 1279px) {*/
-/*    .movies__searchform {*/
-/*        padding-top: 80px;*/
-/*    }*/
-/*}*/
 
 .movies__input {
     all: unset;
