+++ src/components/Profile/Profile.css
@@ -5,7 +5,6 @@
     display: flex;
     flex-direction: column;
     box-sizing: border-box;
-    /*height: 100vh;*/
 }
 
 @media (max-width: 1279px) {
@@ -68,15 +67,6 @@
     justify-content: space-between;
 }
 
-/*.profile__form::after {*/
-/*    content: '-';*/
-/*    height: 1px;*/
-/*    background-color: #424242;*/
-/*    background-color: white;*/
-/*    display: inline;*/
-/*    top: 50%;*/
-/*}*/
-
 .profile__input {
     padding: 0;
     font-size: 11px;
@@ -88,12 +78,8 @@
     text-decoration: none;
     border: none;
     background: transparent;
-    /*box-sizing: border-box;*/
-    /*position: relative;*/
     display: block;
     outline: none;
-    /*font-size: 14px;*/
-    /*line-height: 17px;*/
     color: #FFFFFF;
 }
 
@@ -146,7 +132,6 @@
 }
 
 .profile__save {
-    /*all: unset;*/
     width: 100%;
     background: #4285F4;
     border-radius: 2px;
