--- src/components/SavedMovies/SavedMovies.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const SavedMovies = () => {
-  return <div></div>;
-};
-
-export default SavedMovies;
